+++ src/commands/hardis/scratch/pool/refresh.ts
@@ -1,16 +1,15 @@
 /* jscpd:ignore-start */
 import { spawn } from "child_process";
 import * as c from "chalk";
-import * as fs from "fs-extra";
-import * as path from "path";
 import * as which from "which";
 import { flags, SfdxCommand } from "@salesforce/command";
 import { Messages } from "@salesforce/core";
 import { AnyJson } from "@salesforce/ts-types";
 import { addScratchOrgToPool, getPoolStorage, setPoolStorage } from "../../../../common/utils/poolUtils";
 import { getConfig } from "../../../../config";
-import { createTempDir, execCommand, uxLog } from "../../../../common/utils";
+import { execCommand, uxLog } from "../../../../common/utils";
 import moment = require("moment");
+import { authenticateWithSfdxUrlStore } from "../../../../common/utils/orgUtils";
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const stripAnsi2 = require("strip-ansi");
@@ -51,7 +50,6 @@ export default class ScratchPoolRefresh extends SfdxCommand {
   // Set this to true if your command requires a project workspace; 'requiresProject' is false by default
   protected static requiresProject = true;
 
-  /* jscpd:ignore-end */
   private debugMode = false;
 
   public async run(): Promise<AnyJson> {
@@ -63,6 +61,7 @@ export default class ScratchPoolRefresh extends SfdxCommand {
       uxLog(this, c.yellow("Configuration file must contain a poolConfig property") + "\n" + c.grey(JSON.stringify(config, null, 2)));
       return { outputString: "Configuration file must contain a poolConfig property" };
     }
+
     const maxScratchOrgsNumber = config.poolConfig.maxScratchOrgsNumber || 5;
     uxLog(this, c.grey("Pool config: " + JSON.stringify(config.poolConfig)));
 
@@ -70,6 +69,7 @@ export default class ScratchPoolRefresh extends SfdxCommand {
     const poolStorage = await getPoolStorage({ devHubConn: this.hubOrg.getConnection(), devHubUsername: this.hubOrg.getUsername() });
     let scratchOrgs = poolStorage.scratchOrgs || [];
 
+    /* jscpd:ignore-end */
     // Clean expired orgs
     const minScratchOrgRemainingDays = config.poolConfig.minScratchOrgRemainingDays || 25;
     const scratchOrgsToDelete = [];
@@ -102,12 +102,7 @@ export default class ScratchPoolRefresh extends SfdxCommand {
       await setPoolStorage(poolStorage, { devHubConn: this.hubOrg.getConnection(), devHubUsername: this.hubOrg.getUsername() });
       for (const scratchOrgToDelete of scratchOrgsToDelete) {
         // Authenticate to scratch org to delete
-        const authFile = path.join(await createTempDir(), "sfdxScratchAuth.txt");
-        const authFileContent =
-          scratchOrgToDelete.scratchOrgSfdxAuthUrl || (scratchOrgToDelete.authFileJson ? JSON.stringify(scratchOrgToDelete.authFileJson) : null);
-        await fs.writeFile(authFile, authFileContent, "utf8");
-        const authCommand = `sfdx auth:sfdxurl:store -f ${authFile}`;
-        await execCommand(authCommand, this, { fail: true, output: false });
+        await authenticateWithSfdxUrlStore(scratchOrgToDelete);
         // Delete scratch org
         const deleteCommand = `sfdx force:org:delete --noprompt --targetusername ${scratchOrgToDelete.scratchOrgUsername}`;
         await execCommand(deleteCommand, this, { fail: false, debug: this.debugMode, output: true });