+++ src/common/utils/dataUtils.ts
@@ -11,7 +11,11 @@ export const dataFolderRoot = path.join(".", "scripts", "data");
 // Import data from sfdmu folder
 export async function importData(sfdmuPath: string, commandThis: any, options: any = {}) {
   const dtl = await getDataWorkspaceDetail(sfdmuPath);
+  if (dtl.isDelete === true) {
+    throw new SfdxError("Your export.json contains deletion info, please use appropriate delete command");
+  }
   uxLog(commandThis, c.cyan(`Importing data from ${c.green(dtl.full_label)} ...`));
+  /* jscpd:ignore-start */
   uxLog(commandThis, c.italic(c.grey(dtl.description)));
   const targetUsername = options.targetUsername || commandThis.org.getConnection().username;
   await fs.ensureDir(path.join(sfdmuPath, "logs"));
@@ -23,6 +27,7 @@ export async function importData(sfdmuPath: string, commandThis: any, options: a
     ` -p ${sfdmuPath}` +
     " --noprompt" +
     (config.sfdmuCanModify ? ` --canmodify ${config.sfdmuCanModify}` : "");
+  /* jscpd:ignore-end */
   elapseStart(`import ${dtl.full_label}`);
   await execCommand(dataImportCommand, commandThis, {
     fail: true,
@@ -31,9 +36,41 @@ export async function importData(sfdmuPath: string, commandThis: any, options: a
   elapseEnd(`import ${dtl.full_label}`);
 }
 
+// Delete data using sfdmu folder
+export async function deleteData(sfdmuPath: string, commandThis: any, options: any = {}) {
+  const dtl = await getDataWorkspaceDetail(sfdmuPath);
+  if (dtl.isDelete === false) {
+    throw new SfdxError(
+      "Your export.json does not contain deletion information. Please check http://help.sfdmu.com/full-documentation/advanced-features/delete-from-source"
+    );
+  }
+  uxLog(commandThis, c.cyan(`Deleting data from ${c.green(dtl.full_label)} ...`));
+  uxLog(commandThis, c.italic(c.grey(dtl.description)));
+  const targetUsername = options.targetUsername || commandThis.org.getConnection().username;
+  await fs.ensureDir(path.join(sfdmuPath, "logs"));
+  const config = await getConfig("branch");
+  const dataImportCommand =
+    "sfdx sfdmu:run" +
+    ` --sourceusername ${targetUsername}` +
+    ` -p ${sfdmuPath}` +
+    " --noprompt" +
+    (config.sfdmuCanModify ? ` --canmodify ${config.sfdmuCanModify}` : "");
+  elapseStart(`delete ${dtl.full_label}`);
+  await execCommand(dataImportCommand, commandThis, {
+    fail: true,
+    output: true,
+  });
+  elapseEnd(`delete ${dtl.full_label}`);
+}
+
 // Export data from sfdmu folder
 export async function exportData(sfdmuPath: string, commandThis: any, options: any = {}) {
+  /* jscpd:ignore-start */
   const dtl = await getDataWorkspaceDetail(sfdmuPath);
+  if (dtl.isDelete === true) {
+    throw new SfdxError("Your export.json contains deletion info, please use appropriate delete command");
+  }
+  /* jscpd:ignore-end */
   uxLog(commandThis, c.cyan(`Exporting data from ${c.green(dtl.full_label)} ...`));
   uxLog(commandThis, c.italic(c.grey(dtl.description)));
   const sourceUsername = options.sourceUsername || commandThis.org.getConnection().username;
@@ -93,5 +130,18 @@ export async function getDataWorkspaceDetail(dataWorkspace: string) {
     full_label: `[${folderName}]${folderName != hardisLabel ? `: ${hardisLabel}` : ""}`,
     label: hardisLabel,
     description: hardisDescription,
+    exportJson: exportFileJson,
+    isDelete: isDeleteDataWorkspace(exportFileJson),
   };
+}
+
+// Checks if a sfdmu data workspace can delete object!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!*!*!é!*é!!!é*!é!*!*é!*!*!*!*!*!*!**!!!!!!!*!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!s -> https://help.sfdmu.com/full-documentation/advanced-features/delete-from-source
+export function isDeleteDataWorkspace(exportFileJson: any) {
+  let isDelete = false;
+  for (const objectConfig of exportFileJson.objects) {
+    if (objectConfig?.deleteFromSource === true || objectConfig.operation === "DeleteSource") {
+      isDelete = true;
+    }
+  }
+  return isDelete;
 }